requirements.txt
click==6.7
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
PyYAML==3.12
redis==2.10.5
requests==2.12.5
Werkzeug==0.11.15
